// crates/desktop/src/generated/RepositoryHistory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CommitSummary } from "./CommitSummary";
import type { GitReference } from "./GitReference";

export type RepositoryHistory = { root: string, commits: Array<CommitSummary>, refs: Array<GitReference>, head: string | null, hasMore: boolean, };